automacaoWhatsApp/requirements.txt
fastapi
uvicorn[standard]
openai
supabase
redis
python-dotenv
requests

# uvicorn webhook:app --host 0.0.0.0 --port 8000 --reload
# cloudflared tunnel --url http://localhost:8000
